+++ backend/src/main/java/ee/lio/birthcalculator/service/CalculatorService.java
@@ -2,9 +2,10 @@ package ee.lio.birthcalculator.service;
 
 import ee.lio.birthcalculator.dto.response.CalculationResult;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface CalculatorService {
-    CalculationResult calculate(Double grossSalary,
+    CalculationResult calculate(BigDecimal grossSalary,
                                 LocalDate birthDate);
 }
