+++ backend/src/main/java/ee/lio/birthcalculator/service/SessionService.java
@@ -3,13 +3,14 @@ package ee.lio.birthcalculator.service;
 import ee.lio.birthcalculator.dto.request.BenefitRequest;
 import ee.lio.birthcalculator.model.BenefitSession;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface SessionService {
     String createSession();
 
     void saveSessionData(String sessionId,
-                         Double salary,
+                         BigDecimal salary,
                          LocalDate dob);
 
     BenefitSession getSession(String sessionId);
